requirements.txt
Django==1.10.6
celery==4.0.2
djangorestframework==3.3.2

